+++ RMBTControlServer/src/at/rtr/rmbt/controlServer/SignalResultResource.java
@@ -348,7 +348,7 @@ public class SignalResultResource extends ServerResource {
                                 if (Objects.equals(cell.isActive(), true) &&
                                         cell.getTechnology() != RadioCell.Technology.CONNECTION_WLAN) {
                                     if (locationId == null && !locationIdChanged) {
-                                        locationId = cell.getLocationId();
+                                        locationId = cell.getLocationId().intValue();
                                     } else {
                                         if (!locationIdChanged &&
                                                 !locationId.equals(cell.getLocationId())) {
